[CHECKOUTUI-26] Changing the logic to allow multiple apps to use checkout-ui-settings-server
