+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Fixed
-- Append different files
+- Allow multiple apps to use `checkout-ui-sessints-server`
 ## [0.4.2] - 2021-10-18
 
 ### Added
