+++ manifest.json
@@ -1,7 +1,7 @@
 {
   "name": "checkout-ui-settings-server",
   "vendor": "vtex",
-  "version": "0.4.3-beta.0",
+  "version": "0.4.2",
   "title": "Checkout UI Settings Server",
   "description": "Server custom settings for checkout UI to portal",
   "builders": {
